SOZO-144 | move constants and functions for less redundancy
