+++ hlc/config/constants.py
@@ -1,0 +1,21 @@
+from enum import Enum
+
+
+FORWARD = (0, 1)
+RIGHT_FORWARD = (1, 1)
+RIGHT = (1, 0)
+RIGHT_BACKWARD = (1, -1)
+BACKWARD = (0, -1)
+LEFT_BACKWARD = (-1, -1)
+LEFT = (-1, 0)
+LEFT_FORWARD = (-1, 1)
+
+MOVE_FORWARD = 'move forward'
+TURN_RIGHT = 'turn right 90 degrees'
+
+
+class Orientation(Enum):
+    FORWARD = (0, 1)
+    RIGHT = (1, 0)
+    BACKWARD = (0, -1)
+    LEFT = (-1, 0)
